Move mock_stocks and mock_article_data to global conftest, leave some TODO notes :]
